+++ reacttest/src/redux/forms.js
@@ -1,5 +1,6 @@
 export const InitialFeedback={
 
+    id:'',
     title:'',
     body:''
